+++ invisible-friend-tool/src/class/ClaseParticipante.js
