+++ invisible-friend-tool/src/scripts/Barajar.js
