+++ invisible-friend-tool/src/scripts/EnviarAServidor.js
