+++ invisible-friend-tool/src/scripts/FuncionesUtiles.js
