+++ invisible-friend-tool/src/scripts/ReiniciarTabla.js
