+++ invisible-friend-tool/src/scripts/Variables.js
